Add cargo workspace and split into 2 crates
