+++ Cargo.toml
@@ -1,32 +1,19 @@
-[package]
-name = "crypto_service"
-version = "0.1.0"
-edition = "2021"
-build = "build.rs"
+[workspace]
+resolver = "2"
+members = [
+    "crates/crypto-service-server",
+    "crates/crypto-service-uniffi",
+]
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
+# Shared dependencies that all of the crates need
+[workspace.dependencies]
+clippy = "0.0.302"
 anyhow = "1.0.79"
 axum = { version = "0.7.3", features = ["macros"] }
-clippy = "0.0.302"
-dotenv = "0.15.0"
-reqwest = { version = "0.11.23", features = ["json"] }
 serde = { version = "1.0.194", features = ["derive"] }
 serde_json = "1.0.110"
-tokio = { version = "1.35.1", features = ["full"] }
-uniffi = { version = "0.26.0", features = ["cli"] }
 
-[build-dependencies]
-uniffi = { version = "0.26.0", features = ["build", "bindgen"] }
-uniffi_bindgen = "0.23.0"
-
-[dev-dependencies]
-uniffi = { version = "0.26.0", features = ["bindgen-tests"] }
-
-[lib]
-crate-type = ["staticlib", "cdylib", "lib"]
-
-[[bin]]
-name = "uniffi-bindgen"
-path = "uniffi-bindgen.rs"
+[profile.release]
+incremental = false
+panic = 'unwind'
+codegen-units = 1
