+++ crates/crypto-service-server/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "crypto-service-server"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0.79"
+axum = { version = "0.7.3", features = ["macros"] }
+clippy = "0.0.302"
+dotenv = "0.15.0"
+reqwest = { version = "0.11.23", features = ["json"] }
+serde = { version = "1.0.194", features = ["derive"] }
+serde_json = "1.0.110"
+tokio = { version = "1.35.1", features = ["full"] }
+crypto-service-uniffi = { path = "../crypto-service-uniffi" }
+
+[lib]
+crate-type = ["staticlib", "cdylib", "lib"]
