+++ crates/crypto-service-server/src/api_client/mod.rs
@@ -1,3 +1,2 @@
 pub mod api_client;
-pub mod client_trait;
 pub mod get;
