+++ crates/crypto-service-server/src/binance/mod.rs
@@ -1,0 +1,2 @@
+pub mod binance_client;
+pub mod orderbook_handler;
