+++ crates/crypto-service-server/src/coin_api/assets_handler.rs
@@ -1,10 +1,10 @@
 use axum::extract::{self, Query, State};
+use crypto_service_uniffi::coinapi_service::models::{AssetIcons, AssetIconsParams, SymbolsParams, SymbolsResponse};
 
 use crate::{
-    api_client::api_client::ApiClient, binance_service::binance_client::BinanceClient, coinapi_service::helpers::{AssetIcons, SymbolsResponse}, state::AppState
+    state::AppState
 };
 
-use super::{coinapi_client::CoinApiClient, helpers::{AssetIconsParams, SymbolsParams}};
 
 pub async fn get_asset_icons(
     State(state): extract::State<AppState>,
@@ -32,20 +32,20 @@ pub async fn get_symbols(
         .await
 }
 
-#[uniffi::export]
-pub async fn get_symbols_binding(params: SymbolsParams) -> Vec<SymbolsResponse> {
-    let binance_client: BinanceClient = BinanceClient::new();
-    let coinapi_client: CoinApiClient = CoinApiClient::new();
-    let api_client = ApiClient::new();
+// #[uniffi::export]
+// pub async fn get_symbols_binding(params: SymbolsParams) -> Vec<SymbolsResponse> {
+//     let binance_client: BinanceClient = BinanceClient::new();
+//     let coinapi_client: CoinApiClient = CoinApiClient::new();
+//     let api_client = ApiClient::new();
 
-    let state = AppState::new(binance_client, coinapi_client, api_client);
+//     let state = AppState::new(binance_client, coinapi_client, api_client);
 
-    get_symbols(
-        axum::extract::State(state),
-        Query::from(axum::extract::Query(params)),
-    )
-    .await
-    .map(|x| x.1 .0)
-    .unwrap()
-}
+//     get_symbols(
+//         axum::extract::State(state),
+//         Query::from(axum::extract::Query(params)),
+//     )
+//     .await
+//     .map(|x| x.1 .0)
+//     .unwrap()
+// }
