+++ crates/crypto-service-server/src/coin_api/mod.rs
@@ -1,0 +1,2 @@
+pub mod assets_handler;
+pub mod coinapi_client;
