+++ crates/crypto-service-server/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod state;
+pub mod api_client;
+pub mod binance;
+pub mod coin_api;
