+++ crates/crypto-service-uniffi/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "crypto-service-uniffi"
+version = "0.1.0"
+edition = "2021"
+build = "build.rs"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0.79"
+axum = { version = "0.7.3", features = ["macros"] }
+dotenv = "0.15.0"
+reqwest = { version = "0.11.23", features = ["json"] }
+serde = { version = "1.0.194", features = ["derive"] }
+serde_json = "1.0.110"
+tokio = { version = "1.35.1", features = ["full"] }
+uniffi = { version = "0.26.0", features = ["cli"] }
+
+
+
+[build-dependencies]
+uniffi = { version = "0.26.0", features = ["build", "bindgen"] }
+uniffi_bindgen = "0.23.0"
+
+[dev-dependencies]
+uniffi = { version = "0.26.0", features = ["bindgen-tests"] }
+
+[lib]
+crate-type = ["staticlib", "cdylib", "lib"]
+
+[[bin]]
+name = "uniffi-bindgen"
+path = "uniffi-bindgen.rs"
