+++ crates/crypto-service-uniffi/src/coinapi_service/mod.rs
@@ -1,0 +1,1 @@
+pub mod models;
