+++ crates/crypto-service-uniffi/src/lib.rs
@@ -1,0 +1,5 @@
+pub mod binance_service;
+pub mod coinapi_service;
+pub mod client_trait;
+
+uniffi::include_scaffolding!("crypto_service");
