--- src/binance_service/mod.rs
@@ -1,3 +1,0 @@
-pub mod binance_client;
-pub mod helpers;
-pub mod orderbook_handler;
