--- src/coinapi_service/mod.rs
@@ -1,3 +1,0 @@
-pub mod assets_handler;
-pub mod coinapi_client;
-pub mod helpers;
